Minor refactoring
Too many interfaces for things that wouldn't be / have not need to be,
extended (RadioAddress, PacketHeader, and such). These were refactored
and made into sealed classes.

I also made all classes public, and expose the ICrazyradioDriver
interface allowing any user to create their own ICrazyradioDriver
implementation and easily plug it into the rest of the Crazyflie
framework, or use it in their own.